requirements.txt
Django==3.2.16
djangorestframework==3.14.0
psycopg2==2.8.6
drf-spectacular==0.25.1
environ==1.0
jsonschema==4.17.3
celery[redis]==5.2.7
flower==1.2.0
moviepy==1.0.3
ffmpeg-python==0.2.0
